+++ src/__generated__/ArtistForSaleArtworksGridQuery.graphql.ts
@@ -67,6 +67,7 @@ fragment Artwork_artwork on Artwork {
   is_in_auction
   is_biddable
   is_acquireable
+  is_offerable
   id
   sale {
     is_auction
@@ -186,7 +187,7 @@ return {
   "kind": "Request",
   "operationKind": "query",
   "name": "ArtistForSaleArtworksGridQuery",
-  "id": "a467850c2173b86df3101993985c0f3e",
+  "id": "3ec88fac128c12ebecc80d652f3db6dc",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -347,7 +348,7 @@ return {
                           {
                             "kind": "ScalarField",
                             "alias": null,
-                            "name": "is_biddable",
+                            "name": "is_acquireable",
                             "args": null,
                             "storageKey": null
                           },
@@ -418,11 +419,18 @@ return {
                             "args": null,
                             "storageKey": null
                           },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "is_biddable",
+                            "args": null,
+                            "storageKey": null
+                          },
                           v2,
                           {
                             "kind": "ScalarField",
                             "alias": null,
-                            "name": "is_acquireable",
+                            "name": "is_offerable",
                             "args": null,
                             "storageKey": null
                           },