+++ src/__generated__/FairDetailTestsQuery.graphql.ts
@@ -220,6 +220,7 @@ fragment Artwork_artwork on Artwork {
   is_in_auction
   is_biddable
   is_acquireable
+  is_offerable
   id
   sale {
     is_auction
@@ -522,11 +523,18 @@ v18 = {
 v19 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "is_closed",
+  "name": "is_offerable",
   "args": null,
   "storageKey": null
 },
 v20 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "is_closed",
+  "args": null,
+  "storageKey": null
+},
+v21 = {
   "kind": "LinkedField",
   "alias": null,
   "name": "sale",
@@ -556,7 +564,7 @@ v20 = {
       "args": null,
       "storageKey": null
     },
-    v19,
+    v20,
     {
       "kind": "ScalarField",
       "alias": null,
@@ -567,10 +575,10 @@ v20 = {
     v1
   ]
 },
-v21 = [
+v22 = [
   v8
 ],
-v22 = {
+v23 = {
   "kind": "LinkedField",
   "alias": null,
   "name": "sale_artwork",
@@ -587,7 +595,7 @@ v22 = {
       "args": null,
       "concreteType": "SaleArtworkOpeningBid",
       "plural": false,
-      "selections": v21
+      "selections": v22
     },
     {
       "kind": "LinkedField",
@@ -597,7 +605,7 @@ v22 = {
       "args": null,
       "concreteType": "SaleArtworkCurrentBid",
       "plural": false,
-      "selections": v21
+      "selections": v22
     },
     {
       "kind": "ScalarField",
@@ -615,18 +623,18 @@ v22 = {
       "concreteType": "Sale",
       "plural": false,
       "selections": [
-        v19,
+        v20,
         v1
       ]
     },
     v1
   ]
 },
-v23 = [
+v24 = [
   v5,
   v1
 ],
-v24 = {
+v25 = {
   "kind": "LinkedField",
   "alias": null,
   "name": "artists",
@@ -641,9 +649,9 @@ v24 = {
   ],
   "concreteType": "Artist",
   "plural": true,
-  "selections": v23
+  "selections": v24
 },
-v25 = {
+v26 = {
   "kind": "LinkedField",
   "alias": null,
   "name": "partner",
@@ -651,9 +659,9 @@ v25 = {
   "args": null,
   "concreteType": "Partner",
   "plural": false,
-  "selections": v23
+  "selections": v24
 },
-v26 = [
+v27 = [
   v12,
   v1,
   v13,
@@ -663,54 +671,55 @@ v26 = [
   v17,
   v2,
   v18,
-  v20,
-  v22,
-  v24,
+  v19,
+  v21,
+  v23,
   v25,
+  v26,
   v3
 ],
-v27 = {
+v28 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "__typename",
   "args": null,
   "storageKey": null
 },
-v28 = [
+v29 = [
   v5
 ],
-v29 = {
+v30 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "is_followed",
   "args": null,
   "storageKey": null
 },
-v30 = {
+v31 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "nationality",
   "args": null,
   "storageKey": null
 },
-v31 = {
+v32 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "birthday",
   "args": null,
   "storageKey": null
 },
-v32 = {
+v33 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "deathday",
   "args": null,
   "storageKey": null
 },
-v33 = [
+v34 = [
   v4
 ],
-v34 = {
+v35 = {
   "kind": "LinkedField",
   "alias": null,
   "name": "image",
@@ -718,13 +727,13 @@ v34 = {
   "args": null,
   "concreteType": "Image",
   "plural": false,
-  "selections": v33
+  "selections": v34
 };
 return {
   "kind": "Request",
   "operationKind": "query",
   "name": "FairDetailTestsQuery",
-  "id": "b68a5be52a4ee869ad4c5ac6923491d3",
+  "id": "8c270a1a86b4040f2e4a8be7cc2a7d82",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -1075,7 +1084,7 @@ return {
                         ],
                         "concreteType": "Artwork",
                         "plural": true,
-                        "selections": v26
+                        "selections": v27
                       },
                       v2,
                       {
@@ -1093,7 +1102,7 @@ return {
                         "args": null,
                         "concreteType": "Fair",
                         "plural": false,
-                        "selections": v23
+                        "selections": v24
                       },
                       {
                         "kind": "LinkedField",
@@ -1104,17 +1113,17 @@ return {
                         "concreteType": null,
                         "plural": false,
                         "selections": [
-                          v27,
+                          v28,
                           v1,
                           {
                             "kind": "InlineFragment",
                             "type": "ExternalPartner",
-                            "selections": v28
+                            "selections": v29
                           },
                           {
                             "kind": "InlineFragment",
                             "type": "Partner",
-                            "selections": v28
+                            "selections": v29
                           }
                         ]
                       },
@@ -1152,11 +1161,11 @@ return {
                           v3,
                           v1,
                           v5,
-                          v29,
                           v30,
                           v31,
                           v32,
-                          v34
+                          v33,
+                          v35
                         ]
                       },
                       {
@@ -1187,11 +1196,11 @@ return {
                               v2,
                               v1,
                               v5,
-                              v29,
                               v30,
                               v31,
                               v32,
-                              v34
+                              v33,
+                              v35
                             ]
                           }
                         ]
@@ -1295,21 +1304,22 @@ return {
                                     "concreteType": "Artwork",
                                     "plural": false,
                                     "selections": [
-                                      v12,
+                                      v18,
                                       v2,
                                       v13,
                                       v14,
                                       v15,
                                       v16,
                                       v17,
+                                      v12,
                                       v1,
-                                      v18,
-                                      v20,
-                                      v22,
-                                      v24,
+                                      v19,
+                                      v21,
+                                      v23,
                                       v25,
+                                      v26,
                                       v3,
-                                      v27
+                                      v28
                                     ]
                                   },
                                   v11
@@ -1336,7 +1346,7 @@ return {
                         "args": null,
                         "concreteType": "Image",
                         "plural": false,
-                        "selections": v33
+                        "selections": v34
                       },
                       {
                         "kind": "LinkedField",
@@ -1384,13 +1394,13 @@ return {
                                 "args": null,
                                 "concreteType": "Artwork",
                                 "plural": false,
-                                "selections": v26
+                                "selections": v27
                               }
                             ]
                           }
                         ]
                       },
-                      v27
+                      v28
                     ]
                   }
                 ]