+++ src/__generated__/GeneRefetchQuery.graphql.ts
@@ -99,6 +99,7 @@ fragment Artwork_artwork on Artwork {
   is_in_auction
   is_biddable
   is_acquireable
+  is_offerable
   id
   sale {
     is_auction
@@ -269,7 +270,7 @@ return {
   "kind": "Request",
   "operationKind": "query",
   "name": "GeneRefetchQuery",
-  "id": "022d4fe9decd6be713e0086094e8b529",
+  "id": "16db08a3b713196cf4c3397b3cb110f3",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -567,7 +568,7 @@ return {
                           {
                             "kind": "ScalarField",
                             "alias": null,
-                            "name": "is_biddable",
+                            "name": "is_acquireable",
                             "args": null,
                             "storageKey": null
                           },
@@ -619,11 +620,18 @@ return {
                             "args": null,
                             "storageKey": null
                           },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "is_biddable",
+                            "args": null,
+                            "storageKey": null
+                          },
                           v2,
                           {
                             "kind": "ScalarField",
                             "alias": null,
-                            "name": "is_acquireable",
+                            "name": "is_offerable",
                             "args": null,
                             "storageKey": null
                           },